feat: Add reqCount parameter to useBusiness hook
This commit modifies the useBusiness hook to accept a new customParams property, `reqCount`, which allows requesting the count of businesses. The useBusiness hook is updated to include the `reqCount` parameter in the query parameters when making the API request. This enables fetching the count of businesses along with the business data. The change is implemented in the businesses.ts file.

Also added the logic to edit user and apply pagination