feat: Add support for customParams.type in useBusiness hook
This commit adds support for the `type` property in the `customParams` object of the `useBusiness` hook. It allows specifying the type of businesses to fetch. The `useBusiness` hook is updated to include the `type` parameter in the query parameters when making the API request. This enables fetching businesses of a specific type along with the business data. The change is implemented in the `businesses.ts` file.